% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{coint.criterion}
\alias{coint.criterion}
\title{Cointegration criterion of Athanasopoulos et al. (2016)}
\usage{
coint.criterion(input, detrend = TRUE)
}
\arguments{
\item{input}{a matrix of time series data (the bigger dimension is considered to be the sample size)}

\item{detrend}{whether the data should be detrended, TRUE is recommended}
}
\value{
rank of cointegration as numeric
}
\description{
an information criterion to determine the cointegration rank of a data set
}
