% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_aggregation.R
\name{ecvarma.agg.con}
\alias{ecvarma.agg.con}
\title{Method for contemporaneous aggregation}
\usage{
ecvarma.agg.con(ecvarma, l, Fmat = NULL)
}
\arguments{
\item{ecvarma}{an ECVARMA object, created by ecvarma.define, an ECVARMA aggregation function or an estimated ECVARMA process}

\item{l}{the number of omitted variables}

\item{Fmat}{a matrix to change order of variables or create contemporaneous averages, must have full rank (k x k)}
}
\value{
an ECVARMA object of the aggregated model
}
\description{
A function to derive the contemporaneously aggregated error correction model.
}
