% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_marcelino.R
\name{ecvarma.agg.M}
\alias{ecvarma.agg.M}
\title{temporal aggregation with Marcelino's method (1999)}
\usage{
ecvarma.agg.M(ecvarma, l, f = c("avg", "bop"), start.values = "auto")
}
\arguments{
\item{ecvarma}{an ECVARMA object, created by ecvarma.define, an ECVARMA aggregation function or an estimated ECVARMA process}

\item{l}{the aggregation period length}

\item{f}{the used aggregation method: average aggregation - avg, beginning of period sampling - bop}

\item{start.values}{for the Newton method to find the aggregated error variance-covariance matrix and MA parameter (k x k+k*q_agg), with "auto" these are generated automatically}
}
\value{
an ECVARMA object of the aggregated model
}
\description{
A function to derive the temporally aggregated error correction model.
}
