% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecvarma.R
\name{ecvarma}
\alias{ecvarma}
\title{ECVARMA estimation with restrictions}
\usage{
ecvarma(
  input,
  p,
  r,
  q,
  n = 5,
  include = c("none", "SRconst", "LRconst"),
  R,
  R0 = NULL,
  initial_values = "auto",
  h = "auto",
  Newton = TRUE,
  p.values = TRUE
)
}
\arguments{
\item{input}{a matrix of time series data in levels (the bigger dimension is considered to be t)}

\item{p}{number of lagged first differences}

\item{r}{number of cointegrating vectors, only for r<k, for r=0 use R to fix the alphas to 0}

\item{q}{number of lagged error terms}

\item{n}{number of interations performed with the optimization algorithm}

\item{include}{inclusion of a constant in the short run dynamics, cointegration relationship or not at all}

\item{R}{parameter restriction matrix in \eqn{vec(A_0,\beta_{free},\beta_0,\mu,\alpha,\Gamma,M)=R\delta+R_0}. \eqn{\delta} are the parameters that are actually estimated.}

\item{R0}{a vector of constant values in \eqn{vec(A_0,\beta_{free},\beta_0,\mu,\alpha,\Gamma,M)=R\delta+R_0}}

\item{initial_values}{a vector of first set of all parameter values \eqn{vec(\Sigma,A_0,\beta_{free},\beta_0,\mu,\alpha,\Gamma,M)}, \eqn{\beta_0} and \eqn{\mu} only if a constant in the cointegration relationship or in the short run dynamics is included respectively. With "auto", these are determined automatically, but in rare cases these might not work with the iteration algorithm.}

\item{h}{number of lags in level VAR estimation to determine first values of the error terms, only relevant when initial_values is set to "auto"}

\item{Newton}{option to use either Newton method or nlminb to estimate parameters}

\item{p.values}{whether p-values should be calculated, only possible if last iteration didn't fail}
}
\value{
an ECVARMA object
}
\description{
Estimation algorithm for: \deqn{A_0\Delta Y_t=\alpha\beta^{'} Y_{t-1}+ \mu +\sum_{i=1}^{p} \Gamma _i  \Delta Y_{t-i}+u_t+\sum_{j=1}^{q} M _j  u_{t-j};\; \beta^{'}= (I_r\; \beta_{free}\; \beta_0);\; u_t\sim N(0,\Sigma)}, where \eqn{Y_t} is the time series data in levels, \eqn{\Delta Y_t} the first difference of the time series data and \eqn{u_t} the error term. The cointegration parameters for the first r variables are normalized to an identity matrix. Only \eqn{\beta_{free}} and \eqn{\beta_0} are part of the estimation.
}
\references{
Yap, S. F., & Reinsel, G. C. (1995). Estimation and testing for unit roots in a partially nonstationary vector autoregressive moving average model.  \emph{Journal of the American Statistical Association}, 90(429), 253-267.

Kascha, C., & Trenkler, C. (2015). Simple identification and specification of cointegrated VARMA models. \emph{Journal of Applied Econometrics}, 30(4), 675-702.
}
