% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{ecvarma.sim}
\alias{ecvarma.sim}
\title{Simulation of ECVARMA process}
\usage{
ecvarma.sim(ecvarma, sample.size, n, seed = 0)
}
\arguments{
\item{ecvarma}{an ECVARMA object, created by ecvarma.define, an ECVARMA aggregation function or an estimated ECVARMA process}

\item{sample.size}{final sample size of the simulated data set}

\item{n}{number of omitted initial observations (burn in phase), minimal value is the biggest lag order of the model}

\item{seed}{value for set.seed function}
}
\value{
a matrix of size (T x k)
}
\description{
A function to simulate an ECVARMA process.
}
