% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_fma.R
\name{ecvarma.toFMA}
\alias{ecvarma.toFMA}
\title{Conversion of ECVARMA to FMA specification}
\usage{
ecvarma.toFMA(ecvarma)
}
\arguments{
\item{ecvarma}{an ECVARMA object, created by ecvarma.define, an ECVARMA aggregation function or an estimated ECVARMA process}
}
\value{
an ECVARMA object
}
\description{
Conversion of an ECVARMA model to its FMA specification
}
