% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{lag.select.fma}
\alias{lag.select.fma}
\title{FMA lag criterion by Kascha and Trenkler (2015)}
\usage{
\method{lag}{select.fma}(input, h = "auto", v = 0.5)
}
\arguments{
\item{input}{a matrix of time series data (the bigger dimension is considered to be the sample size)}

\item{h}{an integer that equals the maximum lags tested for the AR and MA dynamic, "auto" uses the recommendation of Kascha and Trenkler}

\item{v}{a parameter for the calculation of the lag criterion, must be greater than 0, Kascha and Trenkler use a value of 0.5}
}
\value{
a list with the recommended number of AR and MA lags, and a matrix with the information criteria values, AR lags = rows (first AR lag reserved for error correction, total number of AR lags is then reduced by one), MA lags = columns
}
\description{
An information criterion to specify the number of AR and MA lags of an (EC)VARMA model in final moving average specification
}
