% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_luetkepohl.R
\name{VECMagg.L}
\alias{VECMagg.L}
\title{Lütkepohl's method for temporal aggregation}
\usage{
VECMagg.L(
  alphas,
  betas,
  AR = NULL,
  l,
  f = c("avg", "bop", "eop", "specify own F-matrix"),
  sig = NULL
)
}
\arguments{
\item{alphas}{a matrix of adjustment parameters (k x r)}

\item{betas}{a matrix of cointegration parameters (r x k) including betas that are normalized to 1 or 0}

\item{AR}{a matrix of the parameters before p lagged first differences (k x kp)}

\item{l}{the aggregation period length}

\item{f}{the used aggregation method: average aggregation - avg, beginning of period sampling - bop, end of period sampling - eop, or an aggregation matrix f (k x kl)}

\item{sig}{the non-aggregated variance-covariance matrix of the error terms}
}
\value{
agg.coeff is a list of the aggregated AR parameters, MA parameters and the variance-covariance matrix of the aggregated error terms
}
\description{
A function to derive the temporally aggregated model using Lütkepohl's method. The aggregated model is in final equation form.
}
